Refactor AppointmentRepositoryImpl and AppointmentsRepository for code consistency and clarity

+++ src/main/kotlin/com/icare/repository/AppointmentsRepository.kt
@@ -5,5 +5,5 @@ import com.icare.model.Appointment
 interface AppointmentsRepository {
     fun insertAppointments(appointment: Appointment): Short
     fun  getPatientAppointments(token: String): List<Appointment>
-    fun getAppointmentsByStatus(status: Int): List<Appointment>
+    fun getAppointmentsByStatus(status: Short): List<Appointment>
 }
